ml/pyproject.toml
[project]
name = "persea-waste-classifier"
version = "1.0.0"
description = "PERSÉA Waste Classifier - YOLOv8 based residue classification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "ultralytics>=8.0.200",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
